+++ tests/test_tier3.py
@@ -2,4 +2,4 @@ import pyclesperanto as cle
 import numpy as np
 
 def histogram():
-    cle.histogram(np.random.rand(30, 30), bins = 250)
+    cle.histogram(np.random.rand(30, 30).astype(int), bins = 250)
